requirements.txt
torch>=2.1
transformers>=4.41
datasets>=2.19
einops>=0.7
scikit-learn>=1.3
scipy>=1.10
numpy>=1.24
tqdm>=4.66
pandas>=2.0
transformer-lens>=1.13.0
faiss-cpu>=1.7.4
pyyaml>=6.0
